+++ services/cdn/src/main/java/com/huaweicloud/sdk/cdn/v1/model/LogObject.java
@@ -45,7 +45,7 @@ public class LogObject  {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value="size")
     
-    private Integer size;
+    private Long size;
 
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -141,7 +141,7 @@ public class LogObject  {
 
     
 
-    public LogObject withSize(Integer size) {
+    public LogObject withSize(Long size) {
         this.size = size;
         return this;
     }
@@ -153,11 +153,11 @@ public class LogObject  {
      * 文件大小(Byte)。
      * @return size
      */
-    public Integer getSize() {
+    public Long getSize() {
         return size;
     }
 
-    public void setSize(Integer size) {
+    public void setSize(Long size) {
         this.size = size;
     }
